refactor(mcp): split Plexus management route
